Refactoring teamMember page
