+++ src/features/TeamMember/index.tsx
@@ -2,11 +2,11 @@ import { PageLayout } from '@/components/Layouts/PageLayout';
 import { Message } from '../Top/Message/Message';
 import { OfficialMedia } from '../Top/OfficialMedia';
 import { OurPartners } from '../Top/OurPartners';
-import { TeamMemberCoreValue } from './components/TeamMemberCoreValue';
 import { Box, Image } from '@chakra-ui/react';
-import { TeamMemberRiders } from './components/TeamMemberRiders';
 import { Title } from '@/components/Elements/Title';
 import { TeamMemberStaff } from './TeamMemberStaff';
+import { TeamMemberCoreValue } from './TeamMemberCoreValue';
+import { TeamMemberRiders } from './TeamMemberRiders';
 
 export const TeamMemberContent = () => {
   return (
